backend/requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google.cloud
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.4
oauthlib==3.2.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
